+++ src/sim_math.hpp
@@ -4,8 +4,8 @@
 #include "fish.hpp"
 namespace SimMath {
     float getDistance(const sf::Vector2f& a, const sf::Vector2f& b);
-    sf::Vector2f polarToCortesian(double rad);
-    double cortesianToPolar(sf::Vector2f vec);
+    sf::Vector2f polarToCortesian(float rad);
+    float cortesianToPolar(sf::Vector2f vec);
     void updatePosition(std::shared_ptr<Fish> fish);
     void separation(std::shared_ptr<Fish> fish, const std::vector<std::shared_ptr<Fish>> fishes);
     void alignment(std::shared_ptr<Fish> fish, const std::vector<std::shared_ptr<Fish>> fishes);
